blog/requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-mdeditor==0.1.16
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.13.1
jdcal==1.4.1
Markdown==3.1.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.1.0
Pygments==2.4.2
pytz==2019.1
PyYAML==5.1.2
six==1.12.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
